test: organize unit tests
