+++ tests/unit/serviceObjects/book/bookDetailsServiceObject.js
