+++ tests/unit/serviceObjects/book/bookServiceObject.js
