+++ tests/unit/serviceObjects/book/ratingServiceObject.js
