+++ tests/unit/serviceObjects/users/bookmarkServiceObject.js
