+++ tests/unit/serviceObjects/users/excerptServiceObject.js
