+++ tests/unit/serviceObjects/users/followServiceObject.js
