+++ tests/unit/users/readLater.test.js
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it } from "vitest"
-import { ReadLaterServices } from "../../src/services/users/readLaterServices.js"
+import { ReadLaterServices } from "../../../src/services/users/readLaterServices.js"
 import { ReadLaterServiceObject } from "../serviceObjects/users/readLaterServiceObject.js"
 
 import dayjs from "dayjs"
